Create an Event for a Group specified by its id pt 1

+++ backend/db/models/event.js
@@ -27,7 +27,7 @@ module.exports = (sequelize, DataTypes) => {
       values: ['In person', 'Online']
     },
     capacity: DataTypes.INTEGER,
-    price: DataTypes.INTEGER,
+    price: DataTypes.DECIMAL,
     startDate: DataTypes.DATE,
     endDate: DataTypes.DATE,
   }, {
